Finished Bar Chart, started Pie Chart
Finished bar chart view. It is not generalized and will only plot average temps (I will go back and correct this if there is time). I have started the pie chart graph. I have planned it such that this pie chart graph can be used with any column from our data set. I have all of the data components in place for this class, now I just have to implement a method that draws the pieChart (aka the hard part).